Split type definitions into separate files
